Ignore :checksum property on updating a locked package.
We are storing :checksum for locked packages (see #3).  This prevents
checking out a branch to be updated.

Also, :checkout option is determined AFTER fetching remote refs for
git sources and `git fetch` instead of `git pull` is used when :checkout
is specified.  This means that specifying `:checkout "master"` does not
checkout the latest worktree of the master branch; it just checks out
the master branch, which is not in sync with the remote branch.  For
these reasons, we try to checkout origin/master instead of the local
master branch.